Price set to int

+++ src/main/java/io/github/sevjet/essensedefence/Wall.java
@@ -2,13 +2,13 @@ package io.github.sevjet.essensedefence;
 
 public class Wall extends Building {
 
-    private double price;
+    private int price;
 
     public Wall() {
-        this(0.0D);
+        this(0);
     }
 
-    public Wall(double price) {
+    public Wall(int price) {
         setHeight(1);
         setWidth(1);
 
@@ -24,7 +24,7 @@ public class Wall extends Building {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
 }
